+++ client/src/apps/match/components/FieldButton.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction } from 'react';
-import { MatchScores } from '../apps/match/MatchApp';
+import { MatchScores } from '../MatchApp';
 import { ScouterPosition } from 'requests';
-import MultiButton from './MultiButton';
+import MultiButton from '../../../components/MultiButton';
 
 type countKeys = keyof MatchScores;
