+++ client/src/apps/pit/components/ImageUploader.tsx
